Show Data Ejemplares

+++ SBAFiliacion/CoreData/Ejemplar+CoreDataProperties.swift
@@ -2,7 +2,7 @@
 //  Ejemplar+CoreDataProperties.swift
 //  SBAFiliacion
 //
-//  Created by Lucas Avellaneda on 23/12/2018.
+//  Created by Lucas Avellaneda on 24/12/2018.
 //  Copyright © 2018 Lucas Avellaneda. All rights reserved.
 //
 //
@@ -22,7 +22,7 @@ extension Ejemplar {
     @NSManaged public var dia: String?
     @NSManaged public var fotos: Int16
     @NSManaged public var haras: Int32
-    @NSManaged public var id: Int32
+    @NSManaged public var id: Int64
     @NSManaged public var madre: String?
     @NSManaged public var manual: Bool
     @NSManaged public var mes: String?
